requirements.txt
h5py
numpy
tqdm
matplotlib
pyyaml
Pillow
opencv-python
moviepy